requirements.txt
absl-py==0.15.0
astunparse==1.6.3
beautifulsoup4==4.10.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
clang==5.0
click==8.0.4
cycler==0.11.0
dataclasses==0.8
dlib==19.23.1
face-detection @ git+https://github.com/hukkelas/DSFD-Pytorch-Inference.git@2bdd997d785e20ea39a911e9b3c451b7cdd3b152
face-recognition==1.3.0
face-recognition-models==0.3.0
facenet-pytorch==2.5.2
filelock==3.4.1
flatbuffers==1.12
gast==0.4.0
gdown==4.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.1.0
idna==3.3
imgviz==1.5.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
imutils==0.5.4
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
labelme==5.0.1
Markdown==3.3.6
matplotlib==3.2.2
natsort==8.1.0
numpy==1.19.5
oauthlib==3.2.0
opencv-contrib-python==3.4.17.63
opencv-python==3.4.17.63
opt-einsum==3.3.0
packaging==21.3
Pillow==8.4.0
plotly==5.7.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==6.0
QtPy==2.0.1
requests==2.27.1
requests-oauthlib==1.3.1
retina-face==0.0.10
rsa==4.8
six==1.15.0
soupsieve==2.3.1
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
torch==1.10.1
torchvision==0.11.2
tqdm==4.63.1
typing-extensions==3.7.4.3
urllib3==1.26.9
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.6.0
